refactor: modularize routes and update docs
